+++ RobotServer/worlds/.frc_rectangle.wbproj
@@ -1,5 +1,5 @@
 Webots Project File version R2022a
-perspectives: 000000ff00000000fd00000003000000000000000000000000fc0100000001fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000000000140000000000000000000000010000011c00000157fc0200000001fb0000001400540065007800740045006400690074006f00720100000016000001570000003f00ffffff00000003000005560000013afc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005560000006900ffffff000004380000015700000001000000020000000100000008fc00000000
+perspectives: 000000ff00000000fd00000003000000000000000000000000fc0100000001fb0000001a0044006f00630075006d0065006e0074006100740069006f006e0000000000000000140000000000000000000000010000011c000001a9fc0200000001fb0000001400540065007800740045006400690074006f00720100000016000001a90000003f00ffffff0000000300000556000000e8fc0100000001fb0000001a0043006f006e0073006f006c00650041006c006c0041006c006c0100000000000005560000006900ffffff00000438000001a900000001000000020000000100000008fc00000000
 simulationViewPerspectives: 000000ff000000010000000200000100000003380100000002010000000100
 sceneTreePerspectives: 000000ff0000000100000002000000c0000000fa0100000002010000000200
 maximizedDockId: -1
